alpha phase is open, this means:
it compiles!
it runs!
it calls without known bug yet!

comments to jolly@eversberg.eu

+++ trace.h
@@ -27,7 +27,7 @@ struct trace {
 	
 	/* type */
 	int category;
-	int serial;
+	unsigned long serial;
 	char name[64];
 
 	/* elements */
